finished data intake logic and cleaned up intake loop as well as whole program, untested

+++ src/ALGOS/RANDOM_FOREST/predict_forest.c
@@ -42,7 +42,7 @@ enum EVENT_TYPE predictForest(random_forest * f, const double * sample){
 
     int vote[NUM_EVENT_TYPE] = {0};
 
-    for(int i = 0; i < f->size; i++){
+    for(size_t i = 0; i < f->size; i++){
 
         // printf("Querying tree %d\n",i+1);
